+++ website/docs/table-group/data-table/examples/table-in-table-with-fixed.jsx
@@ -26,7 +26,7 @@ export default () => {
         </DataTable.Head>
         <DataTable.Body>
           <DataTable.Row tag={RowAccordion}>
-            {(props, row, index, columns) => {
+            {(props, row, index) => {
               return {
                 value: index,
                 active: value.includes(index),
@@ -35,14 +35,10 @@ export default () => {
                     <DataTable data={data}>
                       {/* [1] Set the desired z-index */}
                       <DataTable.Head hidden z-index={1}>
-                        <DataTable.Column
-                          name="keyword"
-                          flex={`0 0 ${columns[0].width}px`}
-                          fixed="left"
-                        />
-                        <DataTable.Column name="kd" flex={`0 0 ${columns[1].width}px`} />
-                        <DataTable.Column name="cpc" flex={`0 0 ${columns[2].width}px`} />
-                        <DataTable.Column name="vol" flex={`0 0 ${columns[3].width}px`} />
+                        <DataTable.Column name="keyword" flex="inherit" fixed="left" />
+                        <DataTable.Column name="kd" flex="inherit" />
+                        <DataTable.Column name="cpc" flex="inherit" />
+                        <DataTable.Column name="vol" flex="inherit" />
                       </DataTable.Head>
                       {/* [2] Set a variable to block the scroll */}
                       <DataTable.Body disabledScroll />
@@ -53,7 +49,7 @@ export default () => {
             }}
           </DataTable.Row>
           <DataTable.Cell name="keyword">
-            {(props, row, index, columns) => {
+            {(props, row, index) => {
               return {
                 children: (
                   <Flex alignItems="center">